refactor(bot-screen): keep only the RFB clipboard header cap
The header check alone removes the unbounded buffering: every other client
message type is fixed-size or bounded by a 16-bit count / one byte, so once
ClientCutText is capped nothing can grow _buf past ~256 KiB + a partial frame.
The feed()/_feed() slicing wrapper was a second layer over the same fact and
its coalesced-frame test exercised behaviour the base parser already had.
Rename the test file into the rfb_filter family.

(cherry picked from commit b9bfa7de4c7e968347923d109c3ae68a66ceb37d)
